feat: Add classified ad message sending and photo upload features with new UI, styling, and localization updates.

+++ classified_ads/views.py
@@ -3,15 +3,14 @@ from django.contrib import messages
 from django.contrib.auth.decorators import login_required
 from django.contrib.auth.mixins import LoginRequiredMixin
 from django.core.mail import send_mail
-from django.http import JsonResponse
-from django.shortcuts import redirect, get_object_or_404
+from django.shortcuts import get_object_or_404, render
 from django.urls import reverse
 from django.utils.translation import gettext as _
 from django.views import generic
-from django.views.decorators.csrf import csrf_exempt
+from django_htmx.http import HttpResponseClientRedirect, HttpResponseClientRefresh
 
 from classified_ads.forms import AdPhotoForm, ClassifiedAdForm, MessageForm
-from cm_main.utils import assert_request_is_ajax, check_edit_permission
+from cm_main.utils import check_edit_permission
 
 from .models import AdPhoto, ClassifiedAd, Categories
 
@@ -48,7 +47,6 @@ class UpdateAdView(LoginRequiredMixin, generic.UpdateView):
     check_edit_permission(self.request, self.get_object().owner)
     context = super().get_context_data(**kwargs)
     context["categories"] = Categories()
-    context["photo_form"] = AdPhotoForm()
     return context
 
   def form_valid(self, form):
@@ -59,12 +57,25 @@ class UpdateAdView(LoginRequiredMixin, generic.UpdateView):
 class DeleteAdView(LoginRequiredMixin, generic.View):
   model = ClassifiedAd
 
+  def get(self, request, pk):
+    ad = get_object_or_404(self.model, pk=pk)
+    return render(
+      request,
+      "cm_main/common/confirm-delete-modal-htmx.html",
+      {
+        "ays_title": _("Classified ads deletion"),
+        "ays_msg": _('Are you sure you want to delete the classified ad "%(title)s"?') % {"title": ad.title},
+        "delete_url": request.get_full_path(),
+        "expected_value": ad.title,
+      },
+    )
+
   def post(self, request, pk):
     ad = get_object_or_404(self.model, pk=pk)
     check_edit_permission(request, ad.owner)
     ad.delete()
     messages.success(request, _('Ad "%(title)s" deleted') % {"title": ad.title})
-    return redirect("classified_ads:list")
+    return HttpResponseClientRedirect(reverse("classified_ads:list"))
 
 
 class ListAdsView(LoginRequiredMixin, generic.ListView):
@@ -79,43 +90,60 @@ class AdDetailView(LoginRequiredMixin, generic.DetailView):
   model = ClassifiedAd
   template_name = "classified_ads/detail.html"
 
-  def get_context_data(self, **kwargs):
-    context = super().get_context_data(**kwargs)
-    context["message_form"] = MessageForm()
-    return context
-
 
 class AdPhotoAddView(LoginRequiredMixin, generic.View):
+  def get(self, request, pk):
+    ad = get_object_or_404(ClassifiedAd, pk=pk)
+    check_edit_permission(request, ad.owner)
+    return render(request, "classified_ads/photo-form.html", {"form": AdPhotoForm(), "ad_id": ad.pk})
+
   def post(self, request, pk):
     form = AdPhotoForm(request.POST, self.request.FILES)
     if form.is_valid():
-      form.instance.ad = get_object_or_404(ClassifiedAd, pk=self.kwargs["pk"])
+      form.instance.ad = get_object_or_404(ClassifiedAd, pk=pk)
       check_edit_permission(request, form.instance.ad.owner)
       form.save()
-      messages.success(self.request, _("Photo added successfully"))
-    return redirect("classified_ads:update", pk=self.kwargs["pk"])
+      # redraw only the image gallery
+      return render(request, "classified_ads/gallery.html", {"edit_gallery": True, "ad": form.instance.ad})
+    else:
+      messages.error(self.request, _("Photo not added: %(errors)s") % {"errors": form.errors})
+      return HttpResponseClientRefresh()
 
 
 @login_required
-@csrf_exempt
 def delete_photo(request, pk):
-  assert_request_is_ajax(request)
   photo = get_object_or_404(AdPhoto, pk=pk)
-  check_edit_permission(request, photo.ad.owner)
-  photo.delete()
-  return JsonResponse({"success": True})
+  ad = photo.ad
+  if request.method == "POST":
+    check_edit_permission(request, photo.ad.owner)
+    photo.delete()
+    # # as the swap is delete below, we don't care of the reponse (but status must be ok)
+    # return HttpResponse(status=200, content="<div>ok</div>")
+    return render(request, "classified_ads/gallery.html", {"edit_gallery": True, "ad": ad})
+  return render(
+    request,
+    "cm_main/common/confirm-delete-modal-htmx.html",
+    {
+      "ays_title": _("Photo deletion"),
+      "ays_msg": _("Are you sure you want to delete this photo?"),
+      "delete_url": request.get_full_path(),
+      # "hx_params": f"hx-target=#photo-{photo.id} hx-swap=delete"
+      "hx_params": "hx-target=#ad-photos hx-swap=outerHTML",
+    },
+  )
 
 
 @login_required
 def send_message(request, pk):
-  # Send a message to the ad owner by email
   ad = get_object_or_404(ClassifiedAd, pk=pk)
-  email = ad.owner.email
-  subject = _("Message from %(username)s about ad %(title)s") % {
-    "username": request.user.full_name,
-    "title": ad.title,
-  }
-  message = _("""Hello %(recipient)s,
+  if request.method == "POST":
+    # Send a message to the ad owner by email
+    email = ad.owner.email
+    subject = _("Message from %(username)s about ad %(title)s") % {
+      "username": request.user.full_name,
+      "title": ad.title,
+    }
+    message = _("""Hello %(recipient)s,
 %(username)s sent you the following message about ad %(title)s:
 ==========
 %(message)s
@@ -126,14 +154,22 @@ The %(site_name)s admin team
 
 %(site_url)s
 """) % {
-    "recipient": ad.owner.full_name,
-    "username": request.user.full_name,
-    "title": ad.title,
-    "message": request.POST.get("message"),
-    "site_name": settings.SITE_NAME,
-    "site_url": settings.SITE_DOMAIN,
-    "email": request.user.email,
-  }
-  send_mail(subject, message, settings.DEFAULT_FROM_EMAIL, [email])
-  messages.success(request, _("Message sent successfully"))
-  return redirect("classified_ads:detail", pk=pk)
+      "recipient": ad.owner.full_name,
+      "username": request.user.full_name,
+      "title": ad.title,
+      "message": request.POST.get("message"),
+      "site_name": settings.SITE_NAME,
+      "site_url": settings.SITE_DOMAIN,
+      "email": request.user.email,
+    }
+    send_mail(subject, message, settings.DEFAULT_FROM_EMAIL, [email], fail_silently=False)
+    messages.success(request, _("Message sent successfully"))
+    return HttpResponseClientRefresh()
+  return render(request, "classified_ads/send-message.html", {"form": MessageForm(), "ad": ad})
+
+
+@login_required
+def get_subcategories(request):
+  category = request.GET.get("category")
+  subcategories = [("", _("Select a subcategory")), *Categories.list_subcategories(category)]
+  return render(request, "classified_ads/form.html#subcategories", {"subcategories": subcategories})
